docs: :memo: Added JSDocs
Added JSDocs to the api to make it easier to understand what is going on in each function.
