+++ apps/api/package.json
@@ -7,17 +7,9 @@
     "deploy": "pnpm build && wrangler deploy",
     "dev": "wrangler dev",
     "test": "vitest run -c vitest.unit.ts",
+    "test:benchmark": "vitest run -c vitest.benchmark.ts",
     "test:integration": "vitest run -c vitest.integration.ts --bail=1",
-    "test:benchmark": "vitest run -c vitest.benchmark.ts"
-  },
-  "devDependencies": {
-    "@cloudflare/vitest-pool-workers": "^0.5.32",
-    "@cloudflare/workers-types": "^4.20241112.0",
-    "@donorbound/tsconfig": "workspace:^",
-    "@vitest/ui": "^2.1.6",
-    "typescript": "^5.7.2",
-    "vitest": "^2.1.6",
-    "wrangler": "^3.91.0"
+    "type-check": "tsc --noEmit"
   },
   "dependencies": {
     "@axiomhq/js": "1.3.0",
@@ -33,5 +25,14 @@
     "hono": "^4.6.12",
     "superjson": "^2.2.1",
     "zod": "^3.23.8"
+  },
+  "devDependencies": {
+    "@cloudflare/vitest-pool-workers": "^0.5.32",
+    "@cloudflare/workers-types": "^4.20241112.0",
+    "@donorbound/tsconfig": "workspace:^",
+    "@vitest/ui": "^2.1.6",
+    "typescript": "^5.7.2",
+    "vitest": "^2.1.6",
+    "wrangler": "^3.91.0"
   }
 }
