+++ apps/api/src/pkgs/environment.ts
@@ -1,9 +1,15 @@
 import { z } from "zod";
 
+/**
+ * Custom Zod schema for validating a Cloudflare ratelimiter object.
+ */
 export const cloudflareRatelimiter = z.custom<{
   limit: (options: { key: string }) => Promise<{ success: boolean }>;
 }>((r) => !!r && typeof r.limit === "function");
 
+/**
+ * Zod schema for validating environment variables.
+ */
 export const zEnvironment = z.object({
   VERSION: z.string().default("unknown"),
   DATABASE_HOST: z.string(),
@@ -28,4 +34,8 @@ export const zEnvironment = z.object({
     }),
 });
 
+/**
+ * Type representing the validated environment variables.
+ * @typedef {z.infer<typeof zEnvironment>} Environment
+ */
 export type Environment = z.infer<typeof zEnvironment>;
