+++ apps/api/src/pkgs/errors/error-codes.ts
@@ -1,0 +1,21 @@
+import { z } from "@hono/zod-openapi";
+
+/**
+ * Enum for error codes.
+ */
+export const ErrorCode = z.enum([
+  "BAD_REQUEST",
+  "FORBIDDEN",
+  "INTERNAL_SERVER_ERROR",
+  "USAGE_EXCEEDED",
+  "DISABLED",
+  "NOT_FOUND",
+  "NOT_UNIQUE",
+  "RATE_LIMITED",
+  "UNAUTHORIZED",
+  "PRECONDITION_FAILED",
+  "INSUFFICIENT_PERMISSIONS",
+  "METHOD_NOT_ALLOWED",
+  "EXPIRED",
+  "DELETE_PROTECTED",
+]);
