+++ apps/api/src/pkgs/metrics/interface.ts
@@ -1,13 +1,18 @@
 import type { Metric } from "@donorbound/metrics";
 
+/**
+ * Interface representing a metrics service.
+ */
 export interface Metrics {
   /**
-   * Emit stores a new metric event
+   * Stores a new metric event.
+   * @param {Metric} metric - The metric event to be stored.
    */
   emit(metric: Metric): void;
 
   /**
-   * flush persists all metrics to durable storage
+   * Persists all metrics to durable storage.
+   * @returns {Promise<void>} A promise that resolves when the operation is complete.
    */
   flush(): Promise<void>;
 }
