+++ apps/api/src/pkgs/metrics/logdrain.ts
@@ -3,11 +3,21 @@ import type { LogSchema } from "@donorbound/logs/types";
 import type { Metric } from "@donorbound/metrics";
 import type { Metrics } from "./interface";
 
+/**
+ * Class implementing the Metrics interface to log metric events.
+ */
 export class LogdrainMetrics implements Metrics {
   private readonly requestId: string;
   private readonly isolateId: string | undefined;
   private readonly environment: LogSchema["environment"];
 
+  /**
+   * Creates an instance of LogdrainMetrics.
+   * @param {Object} options - Options for the LogdrainMetrics.
+   * @param {string} options.requestId - The request ID.
+   * @param {string | undefined} options.isolateId - The isolate ID.
+   * @param {LogSchema["environment"]} options.environment - The environment.
+   */
   constructor(options: {
     requestId: string;
     isolateId: string | undefined;
@@ -18,6 +28,10 @@ export class LogdrainMetrics implements Metrics {
     this.environment = options.environment;
   }
 
+  /**
+   * Emits a new metric event by logging it.
+   * @param {Metric} metric - The metric event to be logged.
+   */
   public emit(metric: Metric): void {
     const log = new Log({
       requestId: this.requestId,
@@ -32,6 +46,11 @@ export class LogdrainMetrics implements Metrics {
     console.info(log.toString());
   }
 
+  /**
+   * Persists all metrics to durable storage.
+   * Currently, this method does nothing.
+   * @returns {Promise<void>} A promise that resolves when the operation is complete.
+   */
   public async flush(): Promise<void> {
     return;
   }
