+++ apps/docs/package.json
@@ -7,7 +7,7 @@
     "build:docs": "node ./scripts/generate-documentation.mjs",
     "dev": "next dev",
     "start": "next start",
-    "types:check": "fumadocs-mdx && tsc --noEmit"
+    "type-check": "fumadocs-mdx && tsc --noEmit"
   },
   "dependencies": {
     "fumadocs-core": "14.5.4",
