+++ internals/id/src/generate.ts
@@ -2,18 +2,29 @@ import baseX from "base-x";
 
 const b58 = baseX("123456789ABCDEFGHJKLMNPQRSTUVWXYZabcdefghijkmnopqrstuvwxyz");
 
+/**
+ * Prefixes for each table ID.
+ * For example, 'd_xxx' will be a donation ID and 'req_xxx' will be a request ID.
+ */
 const prefixes = {
   donation: "d",
   request: "req",
 } as const;
 
+/**
+ * Generates a new ID with a specified prefix.
+ * @template TPrefix
+ * @param {TPrefix} prefix - The prefix for the ID, must be a key of the prefixes object.
+ * @returns {string} A new ID string with the specified prefix.
+ */
 export function newId<TPrefix extends keyof typeof prefixes>(prefix: TPrefix) {
   const buf = crypto.getRandomValues(new Uint8Array(20));
 
   /**
-   * epoch starts more recently so that the 32-bit number space gives a
+   * Epoch starts more recently so that the 32-bit number space gives a
    * significantly higher useful lifetime of around 136 years
    * from 2023-11-14T22:13:20.000Z to 2159-12-22T04:41:36.000Z.
+   * @constant {number}
    */
   const EPOCH_TIMESTAMP = 1_700_000_000_000;
