+++ internals/logs/src/types.ts
@@ -1,4 +1,8 @@
 import type { z } from "zod";
 import type { logSchema } from "./logs";
 
+/**
+ * Type representing the inferred schema of a log entry.
+ * @typedef {z.infer<typeof logSchema>} LogSchema
+ */
 export type LogSchema = z.infer<typeof logSchema>;
